make-meme cmd. finally.
